update on AVirtualButtons and AVirtualJoystick
virtual buttons and joystick will now be able to position themselves on
the bottom of the screen if no x or y is defined on construction. (very
helpful for testing)

Also, using the circularBounds boolean in AVirtualJoystick, you can
decide whether you want your Joystick's knob to be able to be dragged
inside a rectangle (so the joystick axes value can attain extreme values
like [-1;-1]) or inside a circle, which is closer to how a real joystick
would behave.