progressively getting TimeShifter ready for manual speed control
if rewind or replay are now triggered by actions, actions need to be
held for the desired amount of time.
rewind or replay can be triggered by calling the startReplay or
startRewind methods, which accept an optional delay parameter. on delay,
buffer will still occur until delay is reached and replay or rewind is
called and playback is overriden.