TimeShifter : more flexibility - rewind anything
TimeShifter was previously made to record actions only of
NapePhysicsObject - and default ones... which meant that all new objects
to watch needed the same exact parameters or TimeShifter would fail.
It is now possible to record and replay whatever parameter from whatever
object we want using what is (for now) called the "bufferSet":

{object:MyHero , continuous:["x","y","rotation"],
discrete:["health","animationFrame"] }

this is a buffer set. TimeShifter will record MyHero's parameters, and
parameters listed in "continuous" will be interpolated on replay/rewind,
when "discrete" parameters will not.

For example we could now simply record the score of a level simply using
this bufferSet
{object:ScoreKeeper, discrete:"score"}

then just add any bufferSet before it is ever activated with
addBufferSet().

- TimeShifter still remains highly inneficient by recording every
parameters on each frame - even when the parameter didn't change. memory
consumption is horrible, lowering memory by getting rid of uneeded
entries means more 'checks' so less performance (an equilibrium needs to
be found.)