StarlingCamera - part 1
I had to share this for other contributors to look at it , give feedback
and test it - as I feel it is now stable with the zoom feature.

But I still want to fit in the ability to rotate the camera.

and tweak the behavior.

we should try not to use the starling pivot system (as I explain in my
notes) - to port this camera to the spriteview,  and also go well with
the debug views.