StarlingCamera final draft
-zooming and rotation works within bounds.

-zooming can be restricted within bounds (not by default) by using the
restrictZoom property.

-turned target into _target and added _manualPosition for manual
positioning, getters and setters so we can know when we switch to a new
target or not (maybe send a signal or help for camera effects later).

/!\ Please backup your copy of StarlingCamera before trying this out.
it has been tested with and without a target.

demo here : http://labs.gsynuh.com/coffee/LabVieHexEvo_cameradebug2.swf

renderDebug has been left for now.