Cameras, creating less points...
note that MathUtils.rotatePoint changes signature to be more efficient and create less points with the camera. (you have the option to give it a result point that it will set instead of returning a new one on each frame.)
camPos uses that new procedure and is created only once when the camera is set.

 I need confirmation that it works well with SpriteCamera.

I thought I'd add this in as a reminder to myself that there is still too many useless things going on in the cameras (new Point everywhere) , and to get my head back into it as well.