ACitrusCamera - zoomFit returns scale
zoomFit returns its calculated zoom factor.

use case :
-you want to fit 800x600 of your state's space in the screen, so you call zoomFit(800x600)
 if you do zoom() or setZoom() this previous ratio isn't taken into consideration anymore, you actually have  the zoom will be absolute (since baseZoom = 1 by default)

-now if you do the following : _camera.baseZoom = _camera.zoomFit(800x600)
all the following zoom action will be relative to the zoom factor that fits 800x600 of the state in the screen, for example you could now do setZoom(1.2) - and it will actually zoom that 800x600 zone by 1.2...

Why?
camera effects will be similar on every screen dimensions since the zooming will be relative to baseZoom.
and camera effects are now more controlled.

there's also a followTarget boolean which allows to temporarily not follow the target, instead of nulling its reference to it.