@gsynuh two important things:
- fullScreen can't be set up by myself since its value change due to
this function handleStageFullscreen. I don't know the order of the
function but using the AIR simulator it grabs the [SWF.... width &
height instead of the simulator, which is grabbed thanks to
stage.fullScreenWidth...
- Starling's _viewport = _starling.viewPort.clone(); returns me also
the wrong dimension... (in manual mode).

There are 3 call to the resetViewport method, and I can't really get it.