@alamboley - something you might be interested in (untested)
I just thought I'd throw in an idea I've had for a long time :
LoadManager knowing the CitrusObject related to the loaders + sending a signal on every completed load along with the citrus object and art + if said CitrusObject has a handleArtLoaded function, it will call it with the art as argument - cast as ICitrusArt of course cause its the only thing we have in common.

Technically, the idea is simply to know a bit more about loaded stuff - and/or each CitrusObject could setup its own art, touch events and so on, when its own art is loaded.

I thought it wouldn't cost much to add... not sure if it actually makes sense.
don't worry, still working on the rest but I just woke up and couldn't stop myself from proposing this