ACitrusView.as
follow up to
https://github.com/alamboley/Citrus-Engine/commit/1be0e862fd1a4302ca1fd330d6607e788071226b

if the art already has a content it means it doesn't need to be loaded, and if the citrusObject has a handleArtLoaded, we still want to call it with the art in question.

the idea again, is to make a citrusObject able to manage the initialization of its own art (add event listeners to it, scaling it, animation sequence manipulation and so on) instead of having that in the state code.
so if content exists, call handleArtLoaded if it exists...
I suppose the name is not appropriate then ... it might be better to have handleArtReady ?

This would structure the code a bit better in my opinion, and CitrusObject could have this function by default?

I sometimes do myself initialize art in the constructor and so on, but this will help switching from loaded art to already created art depending on what we give to the view parameter as everything would already be in a function that will be called either when view is already a valid content or when view needs to be loaded first.