ACitrusView,
mostly tested now - at least on starling.

fix: try to add object/art to  the load manager if content is null no matter what as even if there are no listeners to the loadmanager's signals, we might still want to call a potential handleArtLoaded on the citrus objects. if the content already exists, which is the case if you set view as a valid content, we won't be adding it to the load manager anyway so no extra cost here.