SpriteArt/StarlingArt, fix potential issue
scenario :
view is a url, the content loads everything is fine.
mid game, view is set to a different url, content will not be null during the new loading process, as such it will not be added to the load manager. so either we null it when view changes, after the usual destroy is called, or we have a flag on the art classes to specify is the current view is loadable/loading.