Added spriteview AnimationSequence.as
- in the fla, no need to stop first frames.
the .fla and .swf for the patch demo are updated to reflect this change.

The MovieClip is wrapped within an AnimationSequence like a starling AnimationSequence.
you can listen for animation complete through onAnimationComplete -
we dynamically add scripts to the MC so it dispatches a signal when one animation ends so we can keep track of them and loop them.

"walk" loops by default.

use SpriteArt.setLoopAnimations() to define other animations that should loop.