new feature: setting the fps for animations in an animation sequence individually
Currently the AnimationSequence-class assumes that the play-speed is
identical for all animations. The new method "setAnimFps" allows to set
this speed individually.

Example: You have an "idle"-animaton with two images each being
displayed for 1 second and you have a "walk"-animation with four images
each being displayed for 0.25 seconds.
Without "setAnimFps" you would have to determine the fastest animation
(in this case "walk") and then you would have add redundant images to
the other animations in order to let them appear to be playing slower
(in this case "idle").