bugfix: AVirtualJoystick - threshold calculation
The threshold property of AVirtualJoystick is supposed to be set between
0 and 1. This means that the normalized distance to the center (which is
also between 0 and 1) has to be used for comparison. But currently the
pixel distance to the center is used instead which will always be >=1
when the knob is not in the center. As a consequence the threshold
setting currently doesn't affect anything.
I've replaced the pixel distance by the normalized distance in order to
fix this.